api/pyproject.toml
[project]
name = "second-brain-api"
version = "0.1.0"
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.0",
  "neo4j==5.23.0",
  "python-dotenv==1.0.1",
  "pydantic==2.7.4"
]


